refactor: NotificationController 네이밍 변경 및 패키지 변경
- NotificationController -> NotificationAPI
